# docs/agent-knowledge-base/hmdp-agent-tools.openapi.yaml
openapi: 3.0.3
info:
  title: HMDP Agent Tools
  version: 0.1.0
  description: >
    黑马点评 AI 导购助手工具定义草案。用于 Dify Custom Tool 或类似 Agent 平台导入。
    真实地址请替换为你的 Spring Boot 服务地址，例如 http://localhost:8081。
servers:
  - url: http://localhost:8081
paths:
...
  /voucher-order/seckill/{id}:
    post:
      operationId: seckillVoucher
      summary: 秒杀优惠券下单
      description: >
        根据优惠券 ID 提交秒杀请求。该接口需要用户登录态，并且 Agent 必须在调用前完成二次确认。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 秒杀券 ID。
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: 用户登录 Token。
      responses:
        "200":
          description: 秒杀下单结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
        "401":
          description: 未登录或登录态失效
components:
  schemas:
    Result:
      type: object
      properties:
        success:
          type: boolean
        errorMsg:
          type: string
          nullable: true
        data:
          nullable: true
        total:
          type: integer
          format: int64
          nullable: true